Week 2 return
Application now displays an image of the map and input is handled by mouse clicks. I think any image can be loaded. It's grayscale values are then used to create a map on which Dijkstra shortest path algoritgm result can be drawn. Pure black is out of bounds, pure white is heaviest weight.